refactor(codex): split handler.ts into handler modules
handler.ts was 1062 lines — handleMessage plus the OAuth-mismatch
recovery ladder, the event/item/tool translators, and the abort
registry. Split into a handler/ folder:
  - state   — in-flight abort registry (getActiveAbort)
  - usage   — CodexUsageExhaustedError + rollout usage-exhausted probe
  - events  — ThreadEvent/ThreadItem → shared stream-state translation
  - message — handleMessage + maybeFallbackForChatGptMismatch (kept
              together since the fallback recurses into handleMessage)

index.ts re-exports handleMessage / getActiveAbort / CodexUsageExhaustedError.
Importers (index, factory, test) updated to the barrel. Verified: codex
handler suite passes (49 tests).
